+++ MekHQ/src/mekhq/gui/campaignOptions/contents/EquipmentAndSuppliesOptionsModel.java
@@ -86,19 +86,6 @@ class EquipmentAndSuppliesOptionsModel {
     boolean variableTechLevel;
     boolean useAmmoByType;
     boolean limitClanTech;
-    String mekWarriorDefaultKit;
-    String vehicleCrewDefaultKit;
-    String aircraftDefaultKit;
-    String mekTechDefaultToolKit;
-    String mechanicDefaultToolKit;
-    String aeroTechDefaultToolKit;
-    String baTechDefaultToolKit;
-    String doctorDefaultToolKit;
-    String adminDefaultToolKit;
-    boolean techsNeedToolKit;
-    boolean addDefaultKitToProcurement;
-    boolean npcFactionArmorKits;
-    boolean requireMekWarriorKitToDeploy;
 
     EquipmentAndSuppliesOptionsModel(@Nonnull CampaignOptions options) {
         acquisitionType = options.get(CampaignOption.ACQUISITIONS_TYPE);
@@ -145,19 +132,6 @@ class EquipmentAndSuppliesOptionsModel {
             index++;
         }
 
-        mekWarriorDefaultKit = options.get(CampaignOption.MEKWARRIOR_DEFAULT_KIT);
-        vehicleCrewDefaultKit = options.get(CampaignOption.VEHICLE_CREW_DEFAULT_KIT);
-        aircraftDefaultKit = options.get(CampaignOption.AIRCRAFT_DEFAULT_KIT);
-        mekTechDefaultToolKit = options.get(CampaignOption.MEK_TECH_DEFAULT_TOOL_KIT);
-        mechanicDefaultToolKit = options.get(CampaignOption.MECHANIC_DEFAULT_TOOL_KIT);
-        aeroTechDefaultToolKit = options.get(CampaignOption.AERO_TECH_DEFAULT_TOOL_KIT);
-        baTechDefaultToolKit = options.get(CampaignOption.BA_TECH_DEFAULT_TOOL_KIT);
-        doctorDefaultToolKit = options.get(CampaignOption.DOCTOR_DEFAULT_TOOL_KIT);
-        adminDefaultToolKit = options.get(CampaignOption.ADMIN_DEFAULT_TOOL_KIT);
-        techsNeedToolKit = options.get(CampaignOption.TECHS_NEED_TOOL_KIT);
-        addDefaultKitToProcurement = options.get(CampaignOption.ADD_DEFAULT_KIT_TO_PROCUREMENT);
-        npcFactionArmorKits = options.get(CampaignOption.NPC_FACTION_ARMOR_KITS);
-        requireMekWarriorKitToDeploy = options.get(CampaignOption.REQUIRE_MEKWARRIOR_KIT_TO_DEPLOY);
         limitByYear = options.get(CampaignOption.LIMIT_BY_YEAR);
         disallowExtinctStuff = options.get(CampaignOption.DISALLOW_EXTINCT_STUFF);
         allowClanPurchases = options.get(CampaignOption.ALLOW_CLAN_PURCHASES);
@@ -173,19 +147,6 @@ class EquipmentAndSuppliesOptionsModel {
     void applyTo(@Nonnull CampaignOptions options) {
         options.set(CampaignOption.ACQUISITIONS_TYPE, acquisitionType);
         options.set(CampaignOption.USE_FUNCTIONAL_APPRAISAL, useFunctionalAppraisal);
-        options.set(CampaignOption.MEKWARRIOR_DEFAULT_KIT, mekWarriorDefaultKit);
-        options.set(CampaignOption.VEHICLE_CREW_DEFAULT_KIT, vehicleCrewDefaultKit);
-        options.set(CampaignOption.AIRCRAFT_DEFAULT_KIT, aircraftDefaultKit);
-        options.set(CampaignOption.MEK_TECH_DEFAULT_TOOL_KIT, mekTechDefaultToolKit);
-        options.set(CampaignOption.MECHANIC_DEFAULT_TOOL_KIT, mechanicDefaultToolKit);
-        options.set(CampaignOption.AERO_TECH_DEFAULT_TOOL_KIT, aeroTechDefaultToolKit);
-        options.set(CampaignOption.BA_TECH_DEFAULT_TOOL_KIT, baTechDefaultToolKit);
-        options.set(CampaignOption.DOCTOR_DEFAULT_TOOL_KIT, doctorDefaultToolKit);
-        options.set(CampaignOption.ADMIN_DEFAULT_TOOL_KIT, adminDefaultToolKit);
-        options.set(CampaignOption.TECHS_NEED_TOOL_KIT, techsNeedToolKit);
-        options.set(CampaignOption.ADD_DEFAULT_KIT_TO_PROCUREMENT, addDefaultKitToProcurement);
-        options.set(CampaignOption.NPC_FACTION_ARMOR_KITS, npcFactionArmorKits);
-        options.set(CampaignOption.REQUIRE_MEKWARRIOR_KIT_TO_DEPLOY, requireMekWarriorKitToDeploy);
         options.set(CampaignOption.ACQUISITION_PERSONNEL_CATEGORY, acquisitionPersonnelCategory);
         options.set(CampaignOption.CLAN_ACQUISITION_PENALTY, clanAcquisitionPenalty);
         options.set(CampaignOption.IS_ACQUISITION_PENALTY, isAcquisitionPenalty);
